position reset button, search and sort options in container
