+++ civictechprojects/static/css/partials/_ProjectSearchContainer.scss
@@ -1,0 +1,4 @@
+.ProjectSearchContainer-reset {
+  display: flex;
+  justify-content: center;
+}
