+++ civictechprojects/static/css/partials/_ProjectSearchSort.scss
@@ -1,0 +1,7 @@
+.ProjectSearchSort-container {
+  display: flex;
+}
+
+.ProjectSearchSort-sortform {
+  padding: 0;
+}
