+++ civictechprojects/static/css/partials/_ResetSearchButton.scss
@@ -1,0 +1,3 @@
+button.reset-search-button:disabled {
+  background-color: $color-grey-disabled;
+}
