+++ civictechprojects/static/css/styles.scss
@@ -38,8 +38,10 @@
 @import 'partials/ProjectCard';
 @import 'partials/ProjectFilterContainer';
 @import 'partials/ProjectSearchBar';
+@import 'partials/ProjectSearchSort';
+@import 'partials/ProjectSearchContainer';
 @import 'partials/ProjectTag';
-@import 'partials/ProjectTagContainer';
+@import 'partials/ResetSearchButton';
 @import 'partials/SectionLink';
 @import 'partials/SubHeader';
 @import 'partials/TagsDisplay';
